Update: 반려견 정보 등록 기능 - 사진을 안 넣어도 등록되게 수정, 예외처리 & 데이터 유효성 검증 추가, Fix: 첨부파일이 없는 반려견은 조회되지 않는 문제 해결, 사진이 없는 반려견의 정보 수정에 실패하는 문제 해결

+++ Goodee-Semi/src/main/java/com/goodee/semi/controller/MyPetUpdateServlet.java
@@ -29,7 +29,7 @@ import jakarta.servlet.http.Part;
 )
 public class MyPetUpdateServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
-    private PetService petService = new PetService();
+    private PetService service = new PetService();
 
     public MyPetUpdateServlet() {
         super();
@@ -111,22 +111,23 @@ public class MyPetUpdateServlet extends HttpServlet {
     				attach.setTypeNo(Attach.PET);
     				attach.setPkNo(pet.getPetNo());
     				
-    				// TODO: 아래의 기능 마저 구현
     				// 기존 이미지파일 삭제
-    				// 1. petNo로 기존 이미지파일의 savedName을 가져옴
-    				// 2. savedName에 해당하는 파일을 저장소에서 삭제
-    				String oldSavedName = petService.selectPetImgSavedName(attach);
+    				// 1) petNo로 기존 이미지파일의 savedName을 가져옴
+    				// 2) savedName에 해당하는 파일을 저장소에서 삭제
+    				String oldSavedName = service.selectPetImgSavedName(attach);
     				System.out.println("[MyPetUpdateServlet] 삭제할 파일명: " + oldSavedName);
-    				File oldSavedFile = new File(AttachService.getUploadDirectory(Attach.PET), oldSavedName);
-    				System.out.println("[MyPetUpdateServlet] 삭제할 파일경로: " + oldSavedFile.getAbsolutePath());
-    				if(oldSavedFile.exists()) {
-    					if (oldSavedFile.delete()) {
-    						System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 삭제에 성공했습니다");
-    					} else {
-    						System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 삭제에 실패했습니다");
-    					}
-    				} else {
-    					System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 파일이 존재하지 않습니다");
+    				if (oldSavedName != null) {
+	    				File oldSavedFile = new File(AttachService.getUploadDirectory(Attach.PET), oldSavedName);
+	    				System.out.println("[MyPetUpdateServlet] 삭제할 파일경로: " + oldSavedFile.getAbsolutePath());
+	    				if(oldSavedFile.exists()) {
+	    					if (oldSavedFile.delete()) {
+	    						System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 삭제에 성공했습니다");
+	    					} else {
+	    						System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 삭제에 실패했습니다");
+	    					}
+	    				} else {
+	    					System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 파일이 존재하지 않습니다");
+	    				}
     				}
     			} catch(Exception e) {
     	            sendErrorResponse("500", "파일 저장/삭제 중 문제가 발생했습니다: " + e.getMessage(), e, response);
@@ -139,7 +140,7 @@ public class MyPetUpdateServlet extends HttpServlet {
 		}
         
         // 3. service 호출
-        int result = petService.updatePet(pet, attach);
+        int result = service.updatePet(pet, attach);
         if (result != 0) {
         	sendSuccessResponse("200", "반려견 정보 수정에 성공했습니다", pet, response);
         	return;
